requirements.txt
Django==1.11.0
djangorestframework==3.7.7
mysqlclient==1.3.10
requests==2.13.0
hashids==1.2.0
